+++ nvml/CHANGELOG.md
@@ -1,5 +1,21 @@
 # CHANGELOG - nvml
 
+## 1.0.5
+
+* Add fan speed to monitored metrics.
+
+## 1.0.4
+
+* Change monotonic_count to gauge for temperature. 
+
+## 1.0.3
+
+* Fix version.
+
+## 1.0.2
+
+* Add GPU temperature metric. 
+
 ## 1.0.1
 
 * [FIXED] Make nvml check quieter on non-GPU hosts (https://github.com/DataDog/integrations-extras/pull/817)